+++ changelog.md
@@ -9,6 +9,7 @@ Codion Change Log
 ### is.codion.swing.framework.ui
 - EntityTablePanel.Config.cellRenderer() and cellEditor() added.
 - EntityComboBox.model() added for consistency.
+- EntityComboBox.createForeignKeyFilterControl() removed.
 
 ## 0.18.18
 ### is.codion.common.db